pyproject.toml
[tool.poetry]
name = "g3-backendpython"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
python-multipart = "^0.0.20"
python-jose = { extras = ["cryptography"], version = "^3.4.0" }
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
